Added functionality to rename a plugin Renaming involves changing the vim_config file to reflect the new name, and renaming the directory into which the Vim plugin was cloned, as well as any directories in the SRC_DIR.
Renaming is planned for use with more elegant automated installations. A repo will be cloned and apt-vim will check the cloned directory for a JSON with @vimpkg. If a JSON is found, its configuration will be used for the install. If no JSON is found, the user will be prompted to enter any required recipes