Added verification of package before update
Now, a package is confirmed to be installed before it is updated. If
the package is not installed, the user will be asked if they would
like to have it installed.

In order for a new package to be installed this way, a valid URL must
be provided (not just a package name, which could be provided for an
update).